refactor: isolate llmfs generation protocol
