+++ crates/agent-engine/src/runtime/agent_loop/namespace_environment/tests/client_io.rs
@@ -1,3 +1,6 @@
+use alan_agent_protocol::Event;
+
+use super::super::generation::read_generation_response_with_text_events;
 use super::*;
 
 #[tokio::test]
